Move renderFiles() to utils and rename to displayNotes().
