+++ src/components/search/search.ts
@@ -2,6 +2,7 @@ import { renderSettingsDropdown } from "../dropdown/dropdown";
 import * as vscode from "vscode";
 import { searchInput } from "./searchInput";
 import { getAllFolderContents } from "../../utils/getAllFolders";
+import { displayNotes } from "../../utils/displayNotes";
 
 export async function search(searchTerm: string, webview: vscode.Webview, context: any) {
 	const styles = webview.asWebviewUri(vscode.Uri.joinPath(context.extensionUri, "src/components/overview", "overview.css"));
@@ -17,7 +18,7 @@ export async function search(searchTerm: string, webview: vscode.Webview, contex
 
 	if (results !== undefined) {
 		folderContentsHTML = await renderFolderContent(results?.folders);
-		fileContentsHTML = await renderFiles(results?.files);
+		fileContentsHTML = await displayNotes(results?.files);
 	} else {
 		folderContentsHTML = `<p>No matching results..</p>`;
 		fileContentsHTML = "";
@@ -187,33 +188,6 @@ async function renderFolderContent(folders: any) {
 	                </div>
 	            </div>
 	            `;
-		})
-		.join("");
-}
-
-async function renderFiles(files: any) {
-	return files
-		.map((file: any) => {
-			const dropdownHtml = renderSettingsDropdown(file);
-			return `
-                <div class="item">
-                    <div class="left file-item" data-folder-name="${file.name}" folder-path="${file.uriPath}">
-                        <p class="folder-name">${file.name}</p>
-                        <p class="mtime">${file.date}</p>
-                    </div>
-
-                    <div class="right">
-                        <div class="settings-container">
-                            <svg xmlns="http://www.w3.org/2000/svg" fill="#fff" height="24" viewBox="0 -960 960 960" width="24">
-                                <path
-                                d="M480.12-139q-34.055 0-57.881-23.803-23.826-23.804-23.826-57.784 0-34.078 23.804-57.952Q446.02-302.413 480-302.413q34.174 0 57.88 23.844 23.707 23.844 23.707 57.881 0 34.036-23.707 57.862Q514.174-139 480.12-139Zm0-259.413q-34.055 0-57.881-23.804Q398.413-446.02 398.413-480q0-34.174 23.804-57.88Q446.02-561.587 480-561.587q34.174 0 57.88 23.707 23.707 23.706 23.707 57.76 0 34.055-23.707 57.881-23.706 23.826-57.76 23.826Zm0-259.174q-34.055 0-57.881-23.894t-23.826-58q0-34.106 23.804-57.813Q446.02-821 480-821q34.174 0 57.88 23.706 23.707 23.707 23.707 57.813t-23.707 58q-23.706 23.894-57.76 23.894Z" />
-                            </svg>
-                            ${dropdownHtml}
-                        </div>
-                    </div>
-
-                </div>
-                `;
 		})
 		.join("");
 }
