+++ src/utils/getLastEditedNotes.ts
@@ -1,7 +1,7 @@
 const fsp = require("fs").promises;
 const path = require('path');
 
-function timeAgo(mtime: EpochTimeStamp) {
+export function timeAgo(mtime: EpochTimeStamp) {
     const currentDate = new Date();
     const pastDate = new Date(mtime);
 
@@ -24,7 +24,7 @@ function timeAgo(mtime: EpochTimeStamp) {
     }
 }
 
-async function readFirstLine(filePath: string): Promise<string> {
+export async function readFirstLine(filePath: string): Promise<string> {
     try {
         const fileContent = await fsp.readFile(filePath, 'utf-8');
         const deltaContent = JSON.parse(fileContent);
